requirements.txt
yt_dlp==2023.07.06
pillow==10.3.0
requests==2.32.3
python-dotenv==1.0.1
openai==1.30.1
gradio==4.44.1

